fix(desktop-update): bundle admission preloads independently
