+++ packages/desktop/update-admission/src/contracts/featureAvailabilityIpc.ts
@@ -1,0 +1,5 @@
+export const desktopFeatureAvailabilityIpcChannels = {
+  changed: "desktop-feature-availability:changed",
+  getSnapshot: "desktop-feature-availability:get-snapshot",
+  isSupported: "desktop-feature-availability:is-supported"
+} as const;
