+++ packages/desktop/update-admission/src/contracts/updateAdmissionIpc.ts
@@ -1,0 +1,9 @@
+export const desktopUpdateAdmissionIpcChannels = {
+  exit: "desktop-update-admission:exit",
+  getState: "desktop-update-admission:get-state",
+  later: "desktop-update-admission:later",
+  manualDownload: "desktop-update-admission:manual-download",
+  retry: "desktop-update-admission:retry",
+  start: "desktop-update-admission:start",
+  state: "desktop-update-admission:state"
+} as const;
