+++ packages/desktop/update-admission/tsup.config.ts
@@ -13,7 +13,9 @@ export default defineConfig({
     "feature-availability/index": "src/feature-availability/index.ts",
     "i18n/index": "src/i18n/index.ts",
     "mandatory-updater/index": "src/mandatory-updater/index.ts",
+    "preload/feature-availability": "src/preload/featureAvailability.ts",
     "preload/index": "src/preload/index.ts",
+    "preload/minimum-version": "src/preload/minimumVersion.ts",
     "react/index": "src/react/index.ts"
   },
   external: ["electron", "react", "react-dom"],
